# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/debts": {
      "get": {
        "operationId": "DebtsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Inclusive start date (shop-local)",
            "schema": {
              "example": "2026-08-01",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "Inclusive end date (shop-local)",
            "schema": {
              "example": "2026-08-31",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "OPEN",
                "PAID",
                "CANCELLED"
              ]
            }
          },
          {
            "name": "overdue",
            "required": false,
            "in": "query",
            "description": "Only OPEN debts past their due date",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Match customer name or phone",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List debts (status/overdue/date/search)",
        "tags": [
          "debts"
        ]
      }
    },
    "/debts/{id}": {
      "get": {
        "operationId": "DebtsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DebtResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get one debt",
        "tags": [
          "debts"
        ]
      },
      "patch": {
        "operationId": "DebtsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDebtDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DebtResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Extend due date / edit note",
        "tags": [
          "debts"
        ]
      }
    },
    "/debts/{id}/pay": {
      "post": {
        "operationId": "DebtsController_pay",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayDebtDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DebtResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Settle a debt in full",
        "tags": [
          "debts"
        ]
      }
    },
    "/expenses": {
      "get": {
        "operationId": "ExpensesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Inclusive start date (shop-local)",
            "schema": {
              "example": "2026-08-01",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "Inclusive end date (shop-local)",
            "schema": {
              "example": "2026-08-31",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Match on note",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List expenses",
        "tags": [
          "expenses"
        ]
      },
      "post": {
        "operationId": "ExpensesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExpenseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create an expense",
        "tags": [
          "expenses"
        ]
      }
    },
    "/expenses/{id}": {
      "patch": {
        "operationId": "ExpensesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExpenseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update an expense",
        "tags": [
          "expenses"
        ]
      },
      "delete": {
        "operationId": "ExpensesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Soft-delete an expense",
        "tags": [
          "expenses"
        ]
      }
    },
...
  },
  "info": {
    "title": "Phone & Accessory Shop API",
    "description": "Backend for a small retail shop: phones, accessories, sales, debts, expenses and statistics. Multi-tenant — all data scoped by shop.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT access token"
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "DebtResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "saleId": {
            "type": "string"
          },
          "saleCode": {
            "type": "string"
          },
          "customerName": {
            "type": "string"
          },
          "customerPhone": {
            "type": "string"
          },
          "productName": {
            "type": "string",
            "description": "Comma-joined product names on the sale"
          },
          "saleTotalAmount": {
            "type": "number",
            "description": "Total amount of the underlying sale"
          },
          "paidAmount": {
            "type": "number",
            "description": "Amount already paid on the sale"
          },
          "amount": {
            "type": "number",
            "description": "Amount still owed"
          },
          "dueDate": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "PAID",
              "CANCELLED"
            ]
          },
          "paidAt": {
            "type": "object",
            "nullable": true
          },
          "note": {
            "type": "object",
            "nullable": true
          },
          "isOverdue": {
            "type": "boolean"
          },
          "daysOverdue": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "saleId",
          "saleCode",
          "customerName",
          "customerPhone",
          "productName",
          "saleTotalAmount",
          "paidAmount",
          "amount",
          "dueDate",
          "status",
          "isOverdue",
          "daysOverdue",
          "createdAt"
        ]
      },
      "PayDebtDto": {
        "type": "object",
        "properties": {
          "paidAt": {
            "type": "string",
            "description": "Settlement date; defaults to today (shop-local)"
          }
        }
      },
      "UpdateDebtDto": {
        "type": "object",
        "properties": {
          "dueDate": {
            "type": "string",
            "description": "Extend the due date"
          },
          "note": {
            "type": "string"
          }
        }
      },
      "CreateExpenseDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "example": 150000
          },
          "note": {
            "type": "string",
            "example": "Ijara"
          },
          "spentAt": {
            "type": "string",
            "description": "Defaults to today (shop-local)"
          }
        },
        "required": [
          "amount",
          "note"
        ]
      },
      "ExpenseResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "note": {
            "type": "string"
          },
          "spentAt": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "amount",
          "note",
          "spentAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateExpenseDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "example": 150000
          },
          "note": {
            "type": "string",
            "example": "Ijara"
          },
          "spentAt": {
            "type": "string",
            "description": "Defaults to today (shop-local)"
          }
        }
      },
...
    }
  }
}